LayoutTests/inspector/console/console-log-proxy-expected.txt

CONSOLE MESSAGE: line 16: [object Proxy]
CONSOLE MESSAGE: line 30: 0
CONSOLE MESSAGE: line 31: [object Proxy]
CONSOLE MESSAGE: line 32: 1
Tests for the console.log with Proxy objects.


== Running test suite: console.log.proxy
-- Running test case: BasicProxyLog
PASS: Logging Proxy objects should not have triggered get trap.

-- Running test case: AvoidTrapWhenCheckingConsoleRepeat
PASS: Logging Proxy objects and primitives should not have triggered get trap.
